Move: files since added new folders
Add feature, render, handler folders
